requirements.txt
asgiref==3.2.10
certifi==2020.6.20
dj-database-url==0.5.0
Django==3.1.2
django-mathfilters==1.0.0
djangorestframework==3.12.1
elasticsearch==7.9.1
elasticsearch-dsl==7.3.0
gunicorn==20.0.4
mysqlclient==2.0.1
psycopg2-binary==2.8.6
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10
whitenoise==5.2.0
